A route now has a capcity
Admin adds capacity along with other route details. Can be viewed from user home page

+++ src/main/java/com/group12/vanPool/data/repository/RouteRepository.java
@@ -3,5 +3,7 @@ package com.group12.vanPool.data.repository;
 import com.group12.vanPool.data.entity.Route;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RouteRepository extends CrudRepository<Route, Long> {
+public interface RouteRepository extends CrudRepository<Route, Integer> {
+
+
 }
